Move AnnotationService class definition from genomicsdb.h to genomicsdb.cc
